MinGW support (GCC 8.x+, with CI support)

+++ serial_port/serial_port.cpp
@@ -55,7 +55,7 @@ bool serial_port::serial_open()
 #ifdef _WIN32
     file_descriptor = CreateFile(port_name, GENERIC_READ | GENERIC_WRITE, 0, NULL, OPEN_EXISTING, 0, NULL);
 
-    if((int)file_descriptor < 0)
+    if((intptr_t)file_descriptor < 0)
     {
         return false;
     }
